Move mock flatmap stuffs to mixin
