+++ src/components/SplitDialog.vue
@@ -205,7 +205,7 @@ export default {
                   ) ||
                   (
                     entry.type === 'Scaffold' && entry.resource &&
-                    (entry.isBodyScaffold || entry.discoverId === "307")
+                    (entry.isBodyScaffold || entry.discoverId === 307)
                   )
                 )
               )
