+++ src/components/viewers/Scaffold.vue
@@ -49,17 +49,14 @@
 
 <script>
 /* eslint-disable no-alert, no-console */
-import { markRaw } from "vue";
 import EventBus from "../EventBus";
 import ContentMixin from "../../mixins/ContentMixin";
 
 import { ScaffoldVuer } from "@abi-software/scaffoldvuer";
-import { getNerveMaps, getTermNerveMaps, getFilterOptions } from "@abi-software/scaffoldvuer/src/scripts/MappedNerves.js";
 import "@abi-software/scaffoldvuer/dist/style.css";
 import { HelpModeDialog } from '@abi-software/map-utilities'
 import '@abi-software/map-utilities/dist/style.css'
-import { FlatmapQueries } from "@abi-software/flatmapvuer/src/services/flatmapQueries.js";
-import { getKnowledgeSource, getReferenceConnectivitiesFromStorage, getReferenceConnectivitiesByAPI } from "@abi-software/flatmapvuer/src/services/flatmapKnowledge.js";
+import { getReferenceConnectivitiesFromStorage, getReferenceConnectivitiesByAPI } from "@abi-software/flatmapvuer/src/services/flatmapKnowledge.js";
 
 export default {
   name: "Scaffold",
@@ -120,108 +117,6 @@ export default {
         EventBus.emit("connectivity-info-close");
       }
     },
-    isViewerMatch: function (entry) {
-      return JSON.stringify(this.entry) === JSON.stringify(entry);
-    },
-    changeConnectivitySource: async function (payload) {
-      const { entry, connectivitySource } = payload;
-      await this.flatmapService.flatmapQueries.queryForConnectivityNew(this.flatmapService.mapImp, entry.featureId[0], connectivitySource);
-      this.tooltipEntry = this.tooltipEntry.map((tooltip) => {
-        if (tooltip.featureId[0] === entry.featureId[0]) {
-          return this.flatmapService.flatmapQueries.updateTooltipData(tooltip);
-        }
-        return tooltip;
-      })
-      EventBus.emit('connectivity-info-open', this.tooltipEntry);
-    },
-    knowledgeTooltipQuery: async function (data) {
-      await this.flatmapService.flatmapQueries.retrieveFlatmapKnowledgeForEvent(this.flatmapService.mapImp, { resource: [data.id] });
-      let tooltip = await this.flatmapService.flatmapQueries.createTooltipData(this.flatmapService.mapImp, {
-        resource: [data.id],
-        label: data.label,
-        provenanceTaxonomy: data.taxons,
-        feature: []
-      })
-      tooltip['knowledgeSource'] = getKnowledgeSource(this.flatmapService.mapImp);
-      tooltip['mapId'] = this.flatmapService.mapImp.provenance.id;
-      tooltip['mapuuid'] = this.flatmapService.mapImp.provenance.uuid;
-      tooltip['nerve-label'] = data['nerve-label'];
-      tooltip['ready'] = true;
-      return tooltip;
-    },
-    getKnowledgeTooltip: async function (payload) {
-      if (this.isViewerMatch(payload.type)) {
-        this.tooltipEntry = [];
-        payload.data.forEach(d => this.tooltipEntry.push({ title: d.label, featureId: [d.id], ready: false }));
-        EventBus.emit('connectivity-info-open', this.tooltipEntry);
-
-        let prom1 = [];
-        // While having placeholders displayed, get details for all paths and then replace.
-        for (let index = 0; index < payload.data.length; index++) {
-          prom1.push(await this.knowledgeTooltipQuery(payload.data[index]));
-        }
-        this.tooltipEntry = await Promise.all(prom1);
-        const featureIds = this.tooltipEntry.map(tooltip => tooltip.featureId[0]);
-        if (featureIds.length > 0) {
-          EventBus.emit('connectivity-info-open', this.tooltipEntry);
-        }
-      }
-    },
-    mockUpFlatmapService: async function() {
-      const flatmapResponse = await fetch(this.flatmapAPI);
-      const flatmapJson = await flatmapResponse.json();
-      const latestFlatmap = flatmapJson
-        .filter(f => f.id === 'human-flatmap_male')
-        .sort((a, b) => b.created.localeCompare(a.created))[0];
-      const flatmapUuid = latestFlatmap.uuid;
-      const flatmapSource = latestFlatmap.sckan['knowledge-source'];
-      const pathwaysResponse = await fetch(`${this.flatmapAPI}/flatmap/${flatmapUuid}/pathways`);
-      const pathwaysJson = await pathwaysResponse.json();
-
-      const flatmapQueries = markRaw(new FlatmapQueries());
-      flatmapQueries.initialise(this.flatmapAPI);
-
-      return {
-        'mockup': true,
-        flatmapQueries: flatmapQueries,
-        getFilterOptions: getFilterOptions,
-        getTermNerveMaps: getTermNerveMaps,
-        'mapImp': {
-          'provenance': {
-            'uuid': flatmapUuid,
-            'connectivity': {
-              ...latestFlatmap.sckan,
-            },
-          },
-          'pathways': pathwaysJson,
-          'resource': this.entry.resource,
-          queryKnowledge : async (keastId) => {
-            const sql = 'select knowledge from knowledge where (source=? or source is null) and entity=? order by source desc';
-            const params = [flatmapSource, keastId];
-            const response = await flatmapQueries.queryKnowledge(sql, params);
-            return JSON.parse(response);
-          },
-          queryLabels : async (entities) => {
-            const sql = `select source, entity, knowledge from knowledge where (source=? or source is null) and entity in (?${', ?'.repeat(entities.length-1)}) order by entity, source desc`;
-            const params = [flatmapSource, ...entities];
-            const response = await flatmapQueries.queryKnowledge(sql, params);
-            const entityLabels = [];
-            let last_entity;
-            for (const row of response) {
-                if (row[1] !== last_entity) {
-                    const knowledge = JSON.parse(row[2]);
-                    entityLabels.push({
-                        entity: row[1],
-                        label: knowledge['label'] || row[1]
-                    })
-                    last_entity = row[1];
-                }
-            }
-            return entityLabels;
-          },
-        },
-      }
-    },
     onResize: function () {
       this.scaffoldCamera.onResize();
     },
@@ -265,14 +160,13 @@ export default {
       }
       this.$refs.scaffold.viewRegion(names);
     },
-    scaffoldIsReady: async function () {
+    scaffoldIsReady: function () {
       this.scaffoldLoaded = true;
       this.$refs.scaffold.$module.graphicsHighlight.highlightColour = [1, 0, 1];
       if (this.isVisible()) {
         let rotation = "free";
         if (this.entry.rotation) rotation = this.entry.rotation;
-        this.flatmapService = await this.mockUpFlatmapService();
-        this.loadConnectivityExplorerConfig(this.flatmapService);
+        this.loadExplorerConfig()
       }
       this.updateViewerSettings();
       EventBus.emit("mapLoaded", this.$refs.scaffold);
@@ -364,8 +258,6 @@ export default {
       apiLocation: process.env.VUE_APP_API_LOCATION,
       scaffoldCamera: undefined,
       scaffoldLoaded: false,
-      flatmapService: undefined,
-      tooltipEntry: [],
       connectivityKnowledge: []
     };
   },
